requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
celery==5.4.0
redis==5.2.1
pgvector==0.3.6
feedparser==6.0.11
python-docx==1.1.2
reportlab==4.2.5
pyyaml==6.0.2
pypdf==5.4.0
langgraph==0.2.60
pytest==8.3.4
httpx==0.28.1
fastembed==0.4.2
playwright==1.48.0
